topic status backend
